requirements.txt
Flask==2.3.3
nltk==3.8.1
google-generativeai==0.4.1
requests>=2.28.0
gunicorn==21.2.0
